Add integration test that runs the app.
Pretty basic test, really: just runs the app using the default
configuration (not that anything is actually configirable yet) and
verifies that the output is kinda sorta okay.

Had to touch a stupid number of lines to accomplish this, though, as:

1. The IT needs to deserialize the JSON output to verify it.
2. In order to reuse those structs, I needed to add a `lib.rs` target.
3. I had to mark those structs as `Deserialize`able, which required
   lifetime/ownership changes, as `&'static str` members can't be
   deserialized.

Anyways, it's done. I think. I'll have to let GH Actions actually run
it, as my Linux box is off getting repaired at the moment.