Cache Docker builds during GH Actions tests.
As part of this, also set the workflow to not run for pull requests, as
this would now present a cache poisoning security risk: don't want to
allow PRs to push into my GH packages registry.

Note: this also required some changes to the HAPI FHIR server (used via
submodule), which are reflected in the new commit of that used here.

This took many, many tries to get right: I've squashed some of the
history here, as it was really pretty ridiculous.