Benchmark POST /Organization ops w/ synthetic data
Most of the work here was on or in support of the new
`fhir_bench_orchestrator::test_framework::post_org` module.

This was a rather major effort to add in support for benchmarking the
first complex/"real" FHIR operation. Testing this operation required
synthetic data, which is generated here via Synthea, a health-data
generation tool.

Support was also added for testing at various concurrency levels, as it
became obvious in the development of this PR that that was also a need.

The output format was also modified a solid amount as it became clearer
how things ought to be structured.

Note: There's some commented-out code here that will be removed in a
future commit -- just want to ensure it's in the history.