Use async HTTP requests in benchmarks.
Should allow for much higher throughput, equivalent to or better than
what might be achieved via a greenthreaded benchmarking system.

Note that this change surfaces an issue with HAPI: it tends to deadlock
now with more than 9 concurrent users. Best as I can tell, though,
that's just a problem with HAPI and not specifically related to this
change; I'd likely see it with sync HTTP requests at a high enough
number of concurrent users. I have not attempted to resolve the issue
with HAPI here.

This change also includes some logging improvements.