Collect latency performance data for ops.
The latency data is collected via the `hdrhistogram` crate, and then
represented a few different ways in the JSON: with specific percentiles,
as a "HistoBlob" (serialized `Histogram` struct), and as an embedded
`.hgram` file. This latter format is suitable for graphing via
<http://hdrhistogram.github.com/HdrHistogram/plotFiles.html> and other
plotting tools.

As part of this effort, the operations' resturn data and error handling
was significantly reworked. The state of each operation is now tracked
via a state machine, which is both more robust and also just neat.

One downside of all this, though, is that code maintainability and
readability is really starting to suffer. The overall "run an operation"
code is getting pretty hoary, and has a lot (too much copy-pasting).
This is mostly on purpose, though: I don't want to build out
abstractions for that until I've got one or two more operations
implemented, so that I can ensure I'm building the **right**
abstractions.