Refactor Docker Compose setups for consistency.
This formalizes the pattern of wrapping the `docker compose` command for
each supported FHIR server with a shell script that handles any of the
customizations required.

Should make dealing with the servers a lot simpler, with them all
following the same pattern.

I hope it will also allow future changes that remove a lot of the
`ServerPlugin` and `ServerHandle` code, in favor of a common
implementation.